requirements.txt
shap
numpy
asgiref
pytz
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
sklearn
xgboost
django
djangorestframework
pandas
joblib
pillow
django-crispy-forms
psycopg2-binary
plotly
